WeaponType::smoke: u8 -> i32

+++ core/src/unit.rs
@@ -36,7 +36,7 @@ pub struct WeaponType {
     pub max_air_distance: Option<i32>,
     pub is_inderect: bool,
     pub reaction_fire: bool,
-    pub smoke: Option<u8>, // TODO: u8 -> i32
+    pub smoke: Option<i32>,
 }
 
 #[derive(Clone, Copy, Debug)]
